+++ exporter/exporterhelper/common.go
@@ -52,7 +52,16 @@ type request interface {
 	onError(error) request
 	// Returns the count of spans/metric points or log records.
 	count() int
+	// marshal serializes the current request into a byte stream
+	marshal() ([]byte, error)
+	// onProcessingFinished calls the optional callback function to handle cleanup after all processing is finished
+	onProcessingFinished()
+	// setOnProcessingFinished allows to set an optional callback function to do the cleanup (e.g. remove the item from persistent queue)
+	setOnProcessingFinished(callback func())
 }
+
+// requestUnmarshaler defines a function which takes a byte slice and unmarshals it into a relevant request
+type requestUnmarshaler func([]byte) (request, error)
 
 // requestSender is an abstraction of a sender for a request independent of the type of the data (traces, metrics, logs).
 type requestSender interface {
@@ -61,7 +70,8 @@ type requestSender interface {
 
 // baseRequest is a base implementation for the request.
 type baseRequest struct {
-	ctx context.Context
+	ctx                        context.Context
+	processingFinishedCallback func()
 }
 
 func (req *baseRequest) context() context.Context {
@@ -70,6 +80,16 @@ func (req *baseRequest) context() context.Context {
 
 func (req *baseRequest) setContext(ctx context.Context) {
 	req.ctx = ctx
+}
+
+func (req *baseRequest) setOnProcessingFinished(callback func()) {
+	req.processingFinishedCallback = callback
+}
+
+func (req *baseRequest) onProcessingFinished() {
+	if req.processingFinishedCallback != nil {
+		req.processingFinishedCallback()
+	}
 }
 
 // baseSettings represents all the options that users can configure.
@@ -159,7 +179,7 @@ type baseExporter struct {
 	qrSender *queuedRetrySender
 }
 
-func newBaseExporter(cfg config.Exporter, set component.ExporterCreateSettings, bs *baseSettings) *baseExporter {
+func newBaseExporter(cfg config.Exporter, set component.ExporterCreateSettings, bs *baseSettings, signal config.DataType, reqUnmarshaler requestUnmarshaler) *baseExporter {
 	be := &baseExporter{
 		Component: componenthelper.New(bs.componentOptions...),
 	}
@@ -169,7 +189,7 @@ func newBaseExporter(cfg config.Exporter, set component.ExporterCreateSettings, 
 		ExporterID:             cfg.ID(),
 		ExporterCreateSettings: set,
 	})
-	be.qrSender = newQueuedRetrySender(cfg.ID().String(), bs.QueueSettings, bs.RetrySettings, &timeoutSender{cfg: bs.TimeoutSettings}, set.Logger)
+	be.qrSender = newQueuedRetrySender(cfg.ID(), signal, bs.QueueSettings, bs.RetrySettings, reqUnmarshaler, &timeoutSender{cfg: bs.TimeoutSettings}, set.Logger)
 	be.sender = be.qrSender
 
 	return be
@@ -189,7 +209,7 @@ func (be *baseExporter) Start(ctx context.Context, host component.Host) error {
 	}
 
 	// If no error then start the queuedRetrySender.
-	return be.qrSender.start()
+	return be.qrSender.start(ctx, host)
 }
 
 // Shutdown all senders and exporter and is invoked during service shutdown.