+++ obsreport/obsreport_test.go
@@ -23,8 +23,8 @@ import (
 	"github.com/stretchr/testify/require"
 	"go.opentelemetry.io/otel/attribute"
 	"go.opentelemetry.io/otel/codes"
-	"go.opentelemetry.io/otel/oteltest"
-	"go.opentelemetry.io/otel/trace"
+	sdktrace "go.opentelemetry.io/otel/sdk/trace"
+	"go.opentelemetry.io/otel/sdk/trace/tracetest"
 
 	"go.opentelemetry.io/collector/component/componenttest"
 	"go.opentelemetry.io/collector/config"
@@ -60,8 +60,8 @@ func TestReceiveTraceDataOp(t *testing.T) {
 	defer doneFn()
 
 	set := componenttest.NewNopReceiverCreateSettings()
-	sr := new(oteltest.SpanRecorder)
-	set.TracerProvider = oteltest.NewTracerProvider(oteltest.WithSpanRecorder(sr))
+	sr := new(tracetest.SpanRecorder)
+	set.TracerProvider = sdktrace.NewTracerProvider(sdktrace.WithSpanProcessor(sr))
 
 	parentCtx, parentSpan := set.TracerProvider.Tracer("test").Start(context.Background(), t.Name())
 	defer parentSpan.End()
@@ -81,7 +81,7 @@ func TestReceiveTraceDataOp(t *testing.T) {
 		rec.EndTracesOp(ctx, format, params[i].items, param.err)
 	}
 
-	spans := sr.Completed()
+	spans := sr.Ended()
 	require.Equal(t, len(params), len(spans))
 
 	var acceptedSpans, refusedSpans int
@@ -90,15 +90,15 @@ func TestReceiveTraceDataOp(t *testing.T) {
 		switch params[i].err {
 		case nil:
 			acceptedSpans += params[i].items
-			assert.Equal(t, attribute.Int64Value(int64(params[i].items)), span.Attributes()[obsmetrics.AcceptedSpansKey])
-			assert.Equal(t, attribute.Int64Value(0), span.Attributes()[obsmetrics.RefusedSpansKey])
-			assert.Equal(t, codes.Unset, span.StatusCode())
+			require.Contains(t, span.Attributes(), attribute.KeyValue{Key: obsmetrics.AcceptedSpansKey, Value: attribute.Int64Value(int64(params[i].items))})
+			require.Contains(t, span.Attributes(), attribute.KeyValue{Key: obsmetrics.RefusedSpansKey, Value: attribute.Int64Value(0)})
+			assert.Equal(t, codes.Unset, span.Status().Code)
 		case errFake:
 			refusedSpans += params[i].items
-			assert.Equal(t, attribute.Int64Value(0), span.Attributes()[obsmetrics.AcceptedSpansKey])
-			assert.Equal(t, attribute.Int64Value(int64(params[i].items)), span.Attributes()[obsmetrics.RefusedSpansKey])
-			assert.Equal(t, codes.Error, span.StatusCode())
-			assert.Equal(t, params[i].err.Error(), span.StatusMessage())
+			require.Contains(t, span.Attributes(), attribute.KeyValue{Key: obsmetrics.AcceptedSpansKey, Value: attribute.Int64Value(0)})
+			require.Contains(t, span.Attributes(), attribute.KeyValue{Key: obsmetrics.RefusedSpansKey, Value: attribute.Int64Value(int64(params[i].items))})
+			assert.Equal(t, codes.Error, span.Status().Code)
+			assert.Equal(t, params[i].err.Error(), span.Status().Description)
 		default:
 			t.Fatalf("unexpected param: %v", params[i])
 		}
@@ -112,8 +112,8 @@ func TestReceiveLogsOp(t *testing.T) {
 	defer doneFn()
 
 	set := componenttest.NewNopReceiverCreateSettings()
-	sr := new(oteltest.SpanRecorder)
-	set.TracerProvider = oteltest.NewTracerProvider(oteltest.WithSpanRecorder(sr))
+	sr := new(tracetest.SpanRecorder)
+	set.TracerProvider = sdktrace.NewTracerProvider(sdktrace.WithSpanProcessor(sr))
 
 	parentCtx, parentSpan := set.TracerProvider.Tracer("test").Start(context.Background(), t.Name())
 	defer parentSpan.End()
@@ -133,7 +133,7 @@ func TestReceiveLogsOp(t *testing.T) {
 		rec.EndLogsOp(ctx, format, params[i].items, param.err)
 	}
 
-	spans := sr.Completed()
+	spans := sr.Ended()
 	require.Equal(t, len(params), len(spans))
 
 	var acceptedLogRecords, refusedLogRecords int
@@ -142,15 +142,15 @@ func TestReceiveLogsOp(t *testing.T) {
 		switch params[i].err {
 		case nil:
 			acceptedLogRecords += params[i].items
-			assert.Equal(t, attribute.Int64Value(int64(params[i].items)), span.Attributes()[obsmetrics.AcceptedLogRecordsKey])
-			assert.Equal(t, attribute.Int64Value(0), span.Attributes()[obsmetrics.RefusedLogRecordsKey])
-			assert.Equal(t, codes.Unset, span.StatusCode())
+			require.Contains(t, span.Attributes(), attribute.KeyValue{Key: obsmetrics.AcceptedLogRecordsKey, Value: attribute.Int64Value(int64(params[i].items))})
+			require.Contains(t, span.Attributes(), attribute.KeyValue{Key: obsmetrics.RefusedLogRecordsKey, Value: attribute.Int64Value(0)})
+			assert.Equal(t, codes.Unset, span.Status().Code)
 		case errFake:
 			refusedLogRecords += params[i].items
-			assert.Equal(t, attribute.Int64Value(0), span.Attributes()[obsmetrics.AcceptedLogRecordsKey])
-			assert.Equal(t, attribute.Int64Value(int64(params[i].items)), span.Attributes()[obsmetrics.RefusedLogRecordsKey])
-			assert.Equal(t, codes.Error, span.StatusCode())
-			assert.Equal(t, params[i].err.Error(), span.StatusMessage())
+			require.Contains(t, span.Attributes(), attribute.KeyValue{Key: obsmetrics.AcceptedLogRecordsKey, Value: attribute.Int64Value(0)})
+			require.Contains(t, span.Attributes(), attribute.KeyValue{Key: obsmetrics.RefusedLogRecordsKey, Value: attribute.Int64Value(int64(params[i].items))})
+			assert.Equal(t, codes.Error, span.Status().Code)
+			assert.Equal(t, params[i].err.Error(), span.Status().Description)
 		default:
 			t.Fatalf("unexpected param: %v", params[i])
 		}
@@ -164,8 +164,8 @@ func TestReceiveMetricsOp(t *testing.T) {
 	defer doneFn()
 
 	set := componenttest.NewNopReceiverCreateSettings()
-	sr := new(oteltest.SpanRecorder)
-	set.TracerProvider = oteltest.NewTracerProvider(oteltest.WithSpanRecorder(sr))
+	sr := new(tracetest.SpanRecorder)
+	set.TracerProvider = sdktrace.NewTracerProvider(sdktrace.WithSpanProcessor(sr))
 
 	parentCtx, parentSpan := set.TracerProvider.Tracer("test").Start(context.Background(), t.Name())
 	defer parentSpan.End()
@@ -185,7 +185,7 @@ func TestReceiveMetricsOp(t *testing.T) {
 		rec.EndMetricsOp(ctx, format, params[i].items, param.err)
 	}
 
-	spans := sr.Completed()
+	spans := sr.Ended()
 	require.Equal(t, len(params), len(spans))
 
 	var acceptedMetricPoints, refusedMetricPoints int
@@ -194,15 +194,15 @@ func TestReceiveMetricsOp(t *testing.T) {
 		switch params[i].err {
 		case nil:
 			acceptedMetricPoints += params[i].items
-			assert.Equal(t, attribute.Int64Value(int64(params[i].items)), span.Attributes()[obsmetrics.AcceptedMetricPointsKey])
-			assert.Equal(t, attribute.Int64Value(0), span.Attributes()[obsmetrics.RefusedMetricPointsKey])
-			assert.Equal(t, codes.Unset, span.StatusCode())
+			require.Contains(t, span.Attributes(), attribute.KeyValue{Key: obsmetrics.AcceptedMetricPointsKey, Value: attribute.Int64Value(int64(params[i].items))})
+			require.Contains(t, span.Attributes(), attribute.KeyValue{Key: obsmetrics.RefusedMetricPointsKey, Value: attribute.Int64Value(0)})
+			assert.Equal(t, codes.Unset, span.Status().Code)
 		case errFake:
 			refusedMetricPoints += params[i].items
-			assert.Equal(t, attribute.Int64Value(0), span.Attributes()[obsmetrics.AcceptedMetricPointsKey])
-			assert.Equal(t, attribute.Int64Value(int64(params[i].items)), span.Attributes()[obsmetrics.RefusedMetricPointsKey])
-			assert.Equal(t, codes.Error, span.StatusCode())
-			assert.Equal(t, params[i].err.Error(), span.StatusMessage())
+			require.Contains(t, span.Attributes(), attribute.KeyValue{Key: obsmetrics.AcceptedMetricPointsKey, Value: attribute.Int64Value(0)})
+			require.Contains(t, span.Attributes(), attribute.KeyValue{Key: obsmetrics.RefusedMetricPointsKey, Value: attribute.Int64Value(int64(params[i].items))})
+			assert.Equal(t, codes.Error, span.Status().Code)
+			assert.Equal(t, params[i].err.Error(), span.Status().Description)
 		default:
 			t.Fatalf("unexpected param: %v", params[i])
 		}
@@ -217,13 +217,12 @@ func TestScrapeMetricsDataOp(t *testing.T) {
 	defer doneFn()
 
 	set := componenttest.NewNopReceiverCreateSettings()
-	sr := new(oteltest.SpanRecorder)
-	set.TracerProvider = oteltest.NewTracerProvider(oteltest.WithSpanRecorder(sr))
+	sr := new(tracetest.SpanRecorder)
+	set.TracerProvider = sdktrace.NewTracerProvider(sdktrace.WithSpanProcessor(sr))
 
 	parentCtx, parentSpan := set.TracerProvider.Tracer("test").Start(context.Background(), t.Name())
 	defer parentSpan.End()
 
-	receiverCtx := ScraperContext(parentCtx, receiver, scraper)
 	params := []testParams{
 		{items: 23, err: partialErrFake},
 		{items: 29, err: errFake},
@@ -235,13 +234,12 @@ func TestScrapeMetricsDataOp(t *testing.T) {
 			Scraper:                scraper,
 			ReceiverCreateSettings: set,
 		})
-		ctx := scrp.StartMetricsOp(receiverCtx)
+		ctx := scrp.StartMetricsOp(parentCtx)
 		assert.NotNil(t, ctx)
-
 		scrp.EndMetricsOp(ctx, params[i].items, params[i].err)
 	}
 
-	spans := sr.Completed()
+	spans := sr.Ended()
 	require.Equal(t, len(params), len(spans))
 
 	var scrapedMetricPoints, erroredMetricPoints int
@@ -250,23 +248,23 @@ func TestScrapeMetricsDataOp(t *testing.T) {
 		switch params[i].err {
 		case nil:
 			scrapedMetricPoints += params[i].items
-			assert.Equal(t, attribute.Int64Value(int64(params[i].items)), span.Attributes()[obsmetrics.ScrapedMetricPointsKey])
-			assert.Equal(t, attribute.Int64Value(0), span.Attributes()[obsmetrics.ErroredMetricPointsKey])
-			assert.Equal(t, codes.Unset, span.StatusCode())
+			require.Contains(t, span.Attributes(), attribute.KeyValue{Key: obsmetrics.ScrapedMetricPointsKey, Value: attribute.Int64Value(int64(params[i].items))})
+			require.Contains(t, span.Attributes(), attribute.KeyValue{Key: obsmetrics.ErroredMetricPointsKey, Value: attribute.Int64Value(0)})
+			assert.Equal(t, codes.Unset, span.Status().Code)
 		case errFake:
 			erroredMetricPoints += params[i].items
-			assert.Equal(t, attribute.Int64Value(0), span.Attributes()[obsmetrics.ScrapedMetricPointsKey])
-			assert.Equal(t, attribute.Int64Value(int64(params[i].items)), span.Attributes()[obsmetrics.ErroredMetricPointsKey])
-			assert.Equal(t, codes.Error, span.StatusCode())
-			assert.Equal(t, params[i].err.Error(), span.StatusMessage())
+			require.Contains(t, span.Attributes(), attribute.KeyValue{Key: obsmetrics.ScrapedMetricPointsKey, Value: attribute.Int64Value(0)})
+			require.Contains(t, span.Attributes(), attribute.KeyValue{Key: obsmetrics.ErroredMetricPointsKey, Value: attribute.Int64Value(int64(params[i].items))})
+			assert.Equal(t, codes.Error, span.Status().Code)
+			assert.Equal(t, params[i].err.Error(), span.Status().Description)
 
 		case partialErrFake:
 			scrapedMetricPoints += params[i].items
 			erroredMetricPoints++
-			assert.Equal(t, attribute.Int64Value(int64(params[i].items)), span.Attributes()[obsmetrics.ScrapedMetricPointsKey])
-			assert.Equal(t, attribute.Int64Value(1), span.Attributes()[obsmetrics.ErroredMetricPointsKey])
-			assert.Equal(t, codes.Error, span.StatusCode())
-			assert.Equal(t, params[i].err.Error(), span.StatusMessage())
+			require.Contains(t, span.Attributes(), attribute.KeyValue{Key: obsmetrics.ScrapedMetricPointsKey, Value: attribute.Int64Value(int64(params[i].items))})
+			require.Contains(t, span.Attributes(), attribute.KeyValue{Key: obsmetrics.ErroredMetricPointsKey, Value: attribute.Int64Value(1)})
+			assert.Equal(t, codes.Error, span.Status().Code)
+			assert.Equal(t, params[i].err.Error(), span.Status().Description)
 		default:
 			t.Fatalf("unexpected err param: %v", params[i].err)
 		}
@@ -281,8 +279,8 @@ func TestExportTraceDataOp(t *testing.T) {
 	defer doneFn()
 
 	set := componenttest.NewNopExporterCreateSettings()
-	sr := new(oteltest.SpanRecorder)
-	set.TracerProvider = oteltest.NewTracerProvider(oteltest.WithSpanRecorder(sr))
+	sr := new(tracetest.SpanRecorder)
+	set.TracerProvider = sdktrace.NewTracerProvider(sdktrace.WithSpanProcessor(sr))
 
 	parentCtx, parentSpan := set.TracerProvider.Tracer("test").Start(context.Background(), t.Name())
 	defer parentSpan.End()
@@ -303,7 +301,7 @@ func TestExportTraceDataOp(t *testing.T) {
 		obsrep.EndTracesOp(ctx, params[i].items, params[i].err)
 	}
 
-	spans := sr.Completed()
+	spans := sr.Ended()
 	require.Equal(t, len(params), len(spans))
 
 	var sentSpans, failedToSendSpans int
@@ -312,15 +310,15 @@ func TestExportTraceDataOp(t *testing.T) {
 		switch params[i].err {
 		case nil:
 			sentSpans += params[i].items
-			assert.Equal(t, attribute.Int64Value(int64(params[i].items)), span.Attributes()[obsmetrics.SentSpansKey])
-			assert.Equal(t, attribute.Int64Value(0), span.Attributes()[obsmetrics.FailedToSendSpansKey])
-			assert.Equal(t, codes.Unset, span.StatusCode())
+			require.Contains(t, span.Attributes(), attribute.KeyValue{Key: obsmetrics.SentSpansKey, Value: attribute.Int64Value(int64(params[i].items))})
+			require.Contains(t, span.Attributes(), attribute.KeyValue{Key: obsmetrics.FailedToSendSpansKey, Value: attribute.Int64Value(0)})
+			assert.Equal(t, codes.Unset, span.Status().Code)
 		case errFake:
 			failedToSendSpans += params[i].items
-			assert.Equal(t, attribute.Int64Value(0), span.Attributes()[obsmetrics.SentSpansKey])
-			assert.Equal(t, attribute.Int64Value(int64(params[i].items)), span.Attributes()[obsmetrics.FailedToSendSpansKey])
-			assert.Equal(t, codes.Error, span.StatusCode())
-			assert.Equal(t, params[i].err.Error(), span.StatusMessage())
+			require.Contains(t, span.Attributes(), attribute.KeyValue{Key: obsmetrics.SentSpansKey, Value: attribute.Int64Value(0)})
+			require.Contains(t, span.Attributes(), attribute.KeyValue{Key: obsmetrics.FailedToSendSpansKey, Value: attribute.Int64Value(int64(params[i].items))})
+			assert.Equal(t, codes.Error, span.Status().Code)
+			assert.Equal(t, params[i].err.Error(), span.Status().Description)
 		default:
 			t.Fatalf("unexpected error: %v", params[i].err)
 		}
@@ -335,8 +333,8 @@ func TestExportMetricsOp(t *testing.T) {
 	defer doneFn()
 
 	set := componenttest.NewNopExporterCreateSettings()
-	sr := new(oteltest.SpanRecorder)
-	set.TracerProvider = oteltest.NewTracerProvider(oteltest.WithSpanRecorder(sr))
+	sr := new(tracetest.SpanRecorder)
+	set.TracerProvider = sdktrace.NewTracerProvider(sdktrace.WithSpanProcessor(sr))
 
 	parentCtx, parentSpan := set.TracerProvider.Tracer("test").Start(context.Background(), t.Name())
 	defer parentSpan.End()
@@ -358,7 +356,7 @@ func TestExportMetricsOp(t *testing.T) {
 		obsrep.EndMetricsOp(ctx, params[i].items, params[i].err)
 	}
 
-	spans := sr.Completed()
+	spans := sr.Ended()
 	require.Equal(t, len(params), len(spans))
 
 	var sentMetricPoints, failedToSendMetricPoints int
@@ -367,15 +365,15 @@ func TestExportMetricsOp(t *testing.T) {
 		switch params[i].err {
 		case nil:
 			sentMetricPoints += params[i].items
-			assert.Equal(t, attribute.Int64Value(int64(params[i].items)), span.Attributes()[obsmetrics.SentMetricPointsKey])
-			assert.Equal(t, attribute.Int64Value(0), span.Attributes()[obsmetrics.FailedToSendMetricPointsKey])
-			assert.Equal(t, codes.Unset, span.StatusCode())
+			require.Contains(t, span.Attributes(), attribute.KeyValue{Key: obsmetrics.SentMetricPointsKey, Value: attribute.Int64Value(int64(params[i].items))})
+			require.Contains(t, span.Attributes(), attribute.KeyValue{Key: obsmetrics.FailedToSendMetricPointsKey, Value: attribute.Int64Value(0)})
+			assert.Equal(t, codes.Unset, span.Status().Code)
 		case errFake:
 			failedToSendMetricPoints += params[i].items
-			assert.Equal(t, attribute.Int64Value(0), span.Attributes()[obsmetrics.SentMetricPointsKey])
-			assert.Equal(t, attribute.Int64Value(int64(params[i].items)), span.Attributes()[obsmetrics.FailedToSendMetricPointsKey])
-			assert.Equal(t, codes.Error, span.StatusCode())
-			assert.Equal(t, params[i].err.Error(), span.StatusMessage())
+			require.Contains(t, span.Attributes(), attribute.KeyValue{Key: obsmetrics.SentMetricPointsKey, Value: attribute.Int64Value(0)})
+			require.Contains(t, span.Attributes(), attribute.KeyValue{Key: obsmetrics.FailedToSendMetricPointsKey, Value: attribute.Int64Value(int64(params[i].items))})
+			assert.Equal(t, codes.Error, span.Status().Code)
+			assert.Equal(t, params[i].err.Error(), span.Status().Description)
 		default:
 			t.Fatalf("unexpected error: %v", params[i].err)
 		}
@@ -390,8 +388,8 @@ func TestExportLogsOp(t *testing.T) {
 	defer doneFn()
 
 	set := componenttest.NewNopExporterCreateSettings()
-	sr := new(oteltest.SpanRecorder)
-	set.TracerProvider = oteltest.NewTracerProvider(oteltest.WithSpanRecorder(sr))
+	sr := new(tracetest.SpanRecorder)
+	set.TracerProvider = sdktrace.NewTracerProvider(sdktrace.WithSpanProcessor(sr))
 
 	parentCtx, parentSpan := set.TracerProvider.Tracer("test").Start(context.Background(), t.Name())
 	defer parentSpan.End()
@@ -413,7 +411,7 @@ func TestExportLogsOp(t *testing.T) {
 		obsrep.EndLogsOp(ctx, params[i].items, params[i].err)
 	}
 
-	spans := sr.Completed()
+	spans := sr.Ended()
 	require.Equal(t, len(params), len(spans))
 
 	var sentLogRecords, failedToSendLogRecords int
@@ -422,15 +420,15 @@ func TestExportLogsOp(t *testing.T) {
 		switch params[i].err {
 		case nil:
 			sentLogRecords += params[i].items
-			assert.Equal(t, attribute.Int64Value(int64(params[i].items)), span.Attributes()[obsmetrics.SentLogRecordsKey])
-			assert.Equal(t, attribute.Int64Value(0), span.Attributes()[obsmetrics.FailedToSendLogRecordsKey])
-			assert.Equal(t, codes.Unset, span.StatusCode())
+			require.Contains(t, span.Attributes(), attribute.KeyValue{Key: obsmetrics.SentLogRecordsKey, Value: attribute.Int64Value(int64(params[i].items))})
+			require.Contains(t, span.Attributes(), attribute.KeyValue{Key: obsmetrics.FailedToSendLogRecordsKey, Value: attribute.Int64Value(0)})
+			assert.Equal(t, codes.Unset, span.Status().Code)
 		case errFake:
 			failedToSendLogRecords += params[i].items
-			assert.Equal(t, attribute.Int64Value(0), span.Attributes()[obsmetrics.SentLogRecordsKey])
-			assert.Equal(t, attribute.Int64Value(int64(params[i].items)), span.Attributes()[obsmetrics.FailedToSendLogRecordsKey])
-			assert.Equal(t, codes.Error, span.StatusCode())
-			assert.Equal(t, params[i].err.Error(), span.StatusMessage())
+			require.Contains(t, span.Attributes(), attribute.KeyValue{Key: obsmetrics.SentLogRecordsKey, Value: attribute.Int64Value(0)})
+			require.Contains(t, span.Attributes(), attribute.KeyValue{Key: obsmetrics.FailedToSendLogRecordsKey, Value: attribute.Int64Value(int64(params[i].items))})
+			assert.Equal(t, codes.Error, span.Status().Code)
+			assert.Equal(t, params[i].err.Error(), span.Status().Description)
 		default:
 			t.Fatalf("unexpected error: %v", params[i].err)
 		}
@@ -441,8 +439,8 @@ func TestExportLogsOp(t *testing.T) {
 
 func TestReceiveWithLongLivedCtx(t *testing.T) {
 	set := componenttest.NewNopReceiverCreateSettings()
-	sr := new(oteltest.SpanRecorder)
-	set.TracerProvider = oteltest.NewTracerProvider(oteltest.WithSpanRecorder(sr))
+	sr := new(tracetest.SpanRecorder)
+	set.TracerProvider = sdktrace.NewTracerProvider(sdktrace.WithSpanProcessor(sr))
 
 	longLivedCtx, parentSpan := set.TracerProvider.Tracer("test").Start(context.Background(), t.Name())
 	defer parentSpan.End()
@@ -465,27 +463,27 @@ func TestReceiveWithLongLivedCtx(t *testing.T) {
 		rec.EndTracesOp(ctx, format, params[i].items, params[i].err)
 	}
 
-	spans := sr.Completed()
+	spans := sr.Ended()
 	require.Equal(t, len(params), len(spans))
 
 	for i, span := range spans {
-		assert.Equal(t, trace.SpanID{}, span.ParentSpanID())
+		assert.False(t, span.Parent().IsValid())
 		require.Equal(t, 1, len(span.Links()))
 		link := span.Links()[0]
 		assert.Equal(t, parentSpan.SpanContext().TraceID(), link.SpanContext.TraceID())
 		assert.Equal(t, parentSpan.SpanContext().SpanID(), link.SpanContext.SpanID())
 		assert.Equal(t, "receiver/"+receiver.String()+"/TraceDataReceived", span.Name())
-		assert.Equal(t, attribute.StringValue(transport), span.Attributes()[obsmetrics.TransportKey])
+		require.Contains(t, span.Attributes(), attribute.KeyValue{Key: obsmetrics.TransportKey, Value: attribute.StringValue(transport)})
 		switch params[i].err {
 		case nil:
-			assert.Equal(t, attribute.Int64Value(int64(params[i].items)), span.Attributes()[obsmetrics.AcceptedSpansKey])
-			assert.Equal(t, attribute.Int64Value(0), span.Attributes()[obsmetrics.RefusedSpansKey])
-			assert.Equal(t, codes.Unset, span.StatusCode())
+			require.Contains(t, span.Attributes(), attribute.KeyValue{Key: obsmetrics.AcceptedSpansKey, Value: attribute.Int64Value(int64(params[i].items))})
+			require.Contains(t, span.Attributes(), attribute.KeyValue{Key: obsmetrics.RefusedSpansKey, Value: attribute.Int64Value(0)})
+			assert.Equal(t, codes.Unset, span.Status().Code)
 		case errFake:
-			assert.Equal(t, attribute.Int64Value(0), span.Attributes()[obsmetrics.AcceptedSpansKey])
-			assert.Equal(t, attribute.Int64Value(int64(params[i].items)), span.Attributes()[obsmetrics.RefusedSpansKey])
-			assert.Equal(t, codes.Error, span.StatusCode())
-			assert.Equal(t, params[i].err.Error(), span.StatusMessage())
+			require.Contains(t, span.Attributes(), attribute.KeyValue{Key: obsmetrics.AcceptedSpansKey, Value: attribute.Int64Value(0)})
+			require.Contains(t, span.Attributes(), attribute.KeyValue{Key: obsmetrics.RefusedSpansKey, Value: attribute.Int64Value(int64(params[i].items))})
+			assert.Equal(t, codes.Error, span.Status().Code)
+			assert.Equal(t, params[i].err.Error(), span.Status().Description)
 		default:
 			t.Fatalf("unexpected error: %v", params[i].err)
 		}